mcp_simple_slackbot/requirements.txt
slack_bolt>=1.18.0
slack_sdk>=3.21.0
python-dotenv>=1.0.0
mcp>=1.0.0
mcp-agent>=0.0.16
httpx>=0.24.1
aiohttp>=3.11.13
uvicorn>=0.23.2

# Development dependencies
pyright>=1.1.379
pytest>=8.3.3
ruff>=0.6.9
